#6 [ui] bottom navigation 연결
