+++ app/src/main/java/com/sopt/now/compose/model/BottomNavigationItem.kt
@@ -1,0 +1,8 @@
+package com.sopt.now.compose.model
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class BottomNavigationItem(
+    val icon: ImageVector,
+    val label: String
+)
