+++ app/src/main/java/com/sopt/now/compose/presentation/auth/LoginActivity.kt
@@ -1,10 +1,13 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.presentation.auth
 
 import android.content.Intent
+import android.os.Build
 import android.os.Bundle
 import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -32,16 +35,28 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.sopt.now.compose.Constants.Companion.MAX_ID_LENGTH
-import com.sopt.now.compose.Constants.Companion.MAX_PASSWORD_LENGTH
-import com.sopt.now.compose.Constants.Companion.MBTI_LENGTH
-import com.sopt.now.compose.Constants.Companion.MIN_ID_LENGTH
-import com.sopt.now.compose.Constants.Companion.MIN_PASSWORD_LENGTH
-import com.sopt.now.compose.Constants.Companion.USER_DATA
+import com.sopt.now.compose.R
+import com.sopt.now.compose.model.SignUpData
+import com.sopt.now.compose.presentation.main.MainActivity
+import com.sopt.now.compose.utils.Constants.Companion.USER_DATA
 import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 
-class SignUpActivity : ComponentActivity() {
+class LoginActivity : ComponentActivity() {
     private lateinit var userData: SignUpData
+
+    private val signupLauncher = registerForActivityResult(
+        ActivityResultContracts.StartActivityForResult()
+    ) {
+        if (it.resultCode == RESULT_OK) {
+            userData = getUserData(it)
+        }
+    }
+
+    private fun getUserData(it: ActivityResult): SignUpData {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU)
+            it.data?.getParcelableExtra(USER_DATA, SignUpData::class.java)!!
+        else it.data?.getParcelableExtra(USER_DATA)!!
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -52,73 +67,45 @@ class SignUpActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    showSignup(onSignupBtnClicked = {
-                        userData = SignUpData(it.id, it.password, it.nickname, it.mbti)
-                        if (validateSignup()) {
-                            showToast(R.string.success_signup)
-                            navigateToLogin()
-                            finish()
+                    ShowLogin(
+                        onLoginBtnClicked = { id, password ->
+                            navigateToMain(id, password)
+                        },
+                        onSignupBtnClicked = {
+                            navigateToSignup()
                         }
-                    })
+                    )
                 }
             }
         }
     }
 
-    private fun navigateToLogin() {
-        Intent(this, LoginActivity::class.java).apply {
+    private fun navigateToSignup() =
+        Intent(this, SignUpActivity::class.java).let { signupLauncher.launch(it) }
+
+    private fun navigateToMain() {
+        Intent(this, MainActivity::class.java).apply {
             putExtra(USER_DATA, userData)
-            setResult(RESULT_OK, this)
-        }
+        }.let { startActivity(it) }
     }
 
-    private fun validateSignup(): Boolean =
-        validateId() && validatePassword() && validateNickname() && validateMBTI()
-
-    private fun validateId(): Boolean {
-        require(userData.id.length in MIN_ID_LENGTH..MAX_ID_LENGTH) {
-            showToast(R.string.fail_sign_up_id)
-            return false
-        }
-        return true
+    private fun navigateToMain(id: String, password: String) {
+        if (validateLogin(id, password)) navigateToMain()
+        else Toast.makeText(this, R.string.fail_login, Toast.LENGTH_SHORT).show()
     }
 
-    private fun validatePassword(): Boolean {
-        require(userData.password.length in MIN_PASSWORD_LENGTH..MAX_PASSWORD_LENGTH) {
-            showToast(R.string.fail_sign_up_password)
-            return false
-        }
-        return true
-    }
+    private fun validateLogin(id: String, password: String): Boolean =
+        id == userData.id && password == userData.password
 
-    private fun validateNickname(): Boolean {
-        require(userData.nickname.trim().isNotEmpty()) {
-            showToast(R.string.fail_sign_up_nickname)
-            return false
-        }
-        return true
-    }
-
-    private fun validateMBTI(): Boolean {
-        require(userData.mbti.length == MBTI_LENGTH) {
-            showToast(R.string.fail_sign_up_mbti)
-            return false
-        }
-        return true
-    }
-
-    private fun showToast(message: Int) =
-        Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
 }
 
 @Composable
-fun showSignup(
-    onSignupBtnClicked: (SignUpData) -> Unit
+fun ShowLogin(
+    onLoginBtnClicked: (String, String) -> Unit,
+    onSignupBtnClicked: () -> Unit
 ) {
     var id by remember { mutableStateOf("") }
     var password by remember { mutableStateOf("") }
-    var nickname by remember { mutableStateOf("") }
-    var mbti by remember { mutableStateOf("") }
 
     Column(
         modifier = Modifier
@@ -127,7 +114,7 @@ fun showSignup(
     ) {
         Spacer(modifier = Modifier.height(20.dp))
         Text(
-            text = stringResource(R.string.sign_up),
+            text = stringResource(R.string.welcome_to_sopt),
             fontSize = 30.sp,
             modifier = Modifier
                 .align(Alignment.CenterHorizontally)
@@ -165,54 +152,34 @@ fun showSignup(
             keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.NumberPassword)
         )
 
-        Spacer(modifier = Modifier.size(30.dp))
-
-        Text(
-            text = stringResource(R.string.nickname),
-            fontSize = 20.sp,
-            color = Color.Black
-        )
-        Spacer(modifier = Modifier.size(10.dp))
-        TextField(
-            value = nickname,
-            onValueChange = { nickname = it },
-            placeholder = { Text(text = stringResource(R.string.input_nickname)) },
-            modifier = Modifier.fillMaxWidth()
-        )
-
-        Spacer(modifier = Modifier.size(30.dp))
-
-        Text(
-            text = stringResource(R.string.mbti),
-            fontSize = 20.sp,
-            color = Color.Black
-        )
-        Spacer(modifier = Modifier.size(10.dp))
-        TextField(
-            value = mbti,
-            onValueChange = { mbti = it },
-            placeholder = { Text(text = stringResource(R.string.input_mbti)) },
-            modifier = Modifier.fillMaxWidth()
-        )
-
         Spacer(modifier = Modifier.weight(5f))
 
         Button(
-            onClick = { onSignupBtnClicked(SignUpData(id, password, nickname, mbti)) },
+            onClick = { onLoginBtnClicked(id, password) },
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Text(text = stringResource(R.string.login))
+        }
+
+        Spacer(modifier = Modifier.size(10.dp))
+
+        Button(
+            onClick = onSignupBtnClicked,
             modifier = Modifier
                 .fillMaxWidth()
         ) {
-            Text(
-                text = stringResource(R.string.new_sign_up)
-            )
+            Text(text = stringResource(R.string.new_sign_up))
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun SignupPreview() {
+fun LoginPreview() {
     NOWSOPTAndroidTheme {
-        showSignup(onSignupBtnClicked = { SignUpData -> })
+        ShowLogin(
+            onLoginBtnClicked = { _, _ -> },
+            onSignupBtnClicked = {}
+        )
     }
 }
