+++ app/src/main/java/com/sopt/now/compose/presentation/main/home/HomeScreen.kt
@@ -1,0 +1,9 @@
+package com.sopt.now.compose.presentation.main.home
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun HomeScreen() {
+    Text(text = "Home")
+}
