+++ app/src/main/java/com/sopt/now/compose/presentation/main/mypage/MyPageScreen.kt
@@ -1,0 +1,9 @@
+package com.sopt.now.compose.presentation.main.mypage
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MyPageScreen() {
+    Text(text = "mypage")
+}
