+++ app/src/main/java/com/sopt/now/compose/presentation/main/search/SearchScreen.kt
@@ -1,0 +1,9 @@
+package com.sopt.now.compose.presentation.main.search
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun SearchScreen() {
+    Text(text = "search")
+}
