+++ app/src/main/java/com/sopt/now/compose/utils/Constants.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.utils
 
 class Constants {
     companion object {
